Change price type to double

+++ src/main/java/ru/made/ex/one/BondTrade.java
@@ -2,7 +2,7 @@ package ru.made.ex.one;
 
 public class BondTrade extends BaseTrade {
 
-    public BondTrade(Float price) {
+    public BondTrade(double price) {
         super(price);
     }
 }
