+++ src/main/java/ru/made/ex/one/Trade.java
@@ -1,5 +1,5 @@
 package ru.made.ex.one;
 
 public interface Trade {
-    public Float getPrice();
+    public double getPrice();
 }
